Renamed Dashboard screen to settings screen
