+++ app/src/main/java/com/nnk/appkira/presentation/features/home/screens/settings/widgets/AppsListSettingsDialog.kt
@@ -1,6 +1,5 @@
-package com.nnk.appkira.presentation.features.home.screens.dashboard.widgets
+package com.nnk.appkira.presentation.features.home.screens.settings.widgets
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
